Move test_precompile_consensus_tests to fp_evm
