+++ frame/evm/precompile/blake2/Cargo.toml
@@ -14,11 +14,6 @@ sp-io = { version = "3.0.0", default-features = false, git = "https://github.com
 fp-evm = { version = "0.8.0", default-features = false, path = "../../../../primitives/evm" }
 evm = { version = "0.25.0", default-features = false, features = ["with-codec"] }
 
-[dev-dependencies]
-hex = "0.4.0"
-serde = "1.0"
-serde_json = "1.0"
-
 [features]
 default = ["std"]
 std = [
