+++ frame/evm/precompile/modexp/Cargo.toml
@@ -16,11 +16,6 @@ evm = { version = "0.25.0", default-features = false, features = ["with-codec"] 
 num = { version = "0.3", features = ["alloc"], default-features = false }
 log = "0.4"
 
-[dev-dependencies]
-hex = "0.4.0"
-serde = "1.0"
-serde_json = "1.0"
-
 [features]
 default = ["std"]
 std = [
