+++ frame/evm/src/lib.rs
@@ -56,6 +56,9 @@
 mod tests;
 pub mod runner;
 
+#[cfg(test)]
+pub mod test_utils;
+
 pub use crate::runner::Runner;
 pub use fp_evm::{
 	Account, Log, Vicinity, ExecutionInfo, CallInfo, CreateInfo, Precompile,
