+++ primitives/evm/Cargo.toml
@@ -20,10 +20,15 @@ serde = { version = "1.0.101", optional = true, features = ["derive"] }
 codec = { package = "parity-scale-codec", version = "2.0.0", default-features = false }
 evm = { version = "0.25.0", default-features = false, features = ["with-codec"] }
 impl-trait-for-tuples = "0.1"
+hex = { version = "0.4.0", optional = true }
+serde_json = { version = "1.0", optional = true }
+
 
 [features]
 default = ["std"]
 std = [
+	"hex",
+	"serde_json",
 	"sp-core/std",
 	"sp-std/std",
 	"serde",
